Guess the Number Game (Python3)
This is a simple game I made using Python 3.4. The game asks for your name and for you to guess the number. I learned how to import modules here. This is also an updated version (not much of an update to be honest) where the player is asked if they would like to play again. I may update this even further in the future.